Fixed crash when restoring, but there are problems restoring, especially around creating ports. Might need to go deeo into how this is working and maybe redo some of it

+++ src/net/merayen/merasynth/client/signalgenerator/Net.java
@@ -23,7 +23,7 @@ public class Net extends Node {
 	}
 
 	protected void onRestore(JSONObject state) {
-		frequency = (float)state.get("frequency");
+		frequency = ((Double)state.get("frequency")).floatValue();
 	}
 
 	protected void onReceive(String port, DataPacket dp) {
